Share test suite between verifiers
